Rename marginals plugin into header and footer plugin
