+++ src/marpit.js
@@ -6,8 +6,8 @@ import marpitApplyDirectives from './markdown/directives/apply'
 import marpitBackgroundImage from './markdown/background_image'
 import marpitComment from './markdown/comment'
 import marpitContainerPlugin from './markdown/container'
+import marpitHeaderAndFooter from './markdown/header_and_footer'
 import marpitInlineSVG from './markdown/inline_svg'
-import marpitMarginals from './markdown/marginals'
 import marpitParseDirectives from './markdown/directives/parse'
 import marpitParseImage from './markdown/parse_image'
 import marpitSlide from './markdown/slide'
@@ -90,7 +90,7 @@ class Marpit {
       .use(marpitSlide)
       .use(marpitParseDirectives, this)
       .use(marpitApplyDirectives)
-      .use(marpitMarginals)
+      .use(marpitHeaderAndFooter)
       .use(marpitSlideContainer, this.slideContainers)
       .use(marpitContainerPlugin, this.containers)
       .use(marpitParseImage, { filters: this.options.filters })
